Update index.html
Jacob’s Ladder is as much a discovery as it is a creation. While it will always be the latter because I made it, it is moreso former every time it is experienced it in a new way and from a new perspective.

Experience Jacob’s Ladder open to discovery and perspective. It has no terms that need be met, but the way you approach it will define what you will be capable of seeing. Be curious. Discover what it becomes with time and God. 

It is strange creating something I don’t understand. God is the Great Creator, and all I ostensibly create is merely with the building blocks He has given me that are infinitely more complex than I can imagine. It is a great gift that we can understand so much about the world. Let us not fall into the trap of assuming we know everything about it just because we’re not confused. 

2 Nephi 9:28-29

28 O that cunning plan of the evil one! O the vainness, and the frailties, and the foolishness of men! When they are learned they think they are wise, and they hearken not unto the counsel of God, for they set it aside, supposing they know of themselves, wherefore, their wisdom is foolishness and it profiteth them not. And they shall perish.

29 But to be learned is good if they hearken unto the counsels of God.